ml/pneumonia_detection/pneumonia_detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
streamlit>=1.28.0
opencv-python>=4.8.0
pydicom>=2.4.0
lime>=0.2.0
fpdf2>=2.7.0
plotly>=5.18.0
openai>=1.30.0
python-dotenv>=1.0.0
